Fixed Bugs in authcli update in homepage differernt pages and other useful basic file structure improvement
