+++ src/main/java/com/TextIt/UI/InboxPage.java
@@ -1,0 +1,8 @@
+package com.TextIt.UI;
+
+public class InboxPage {
+
+    public static void main(String[] args) {
+
+    }
+}
