+++ src/main/java/com/TextIt/UI/SearchPage.java
@@ -1,0 +1,7 @@
+package com.TextIt.UI;
+
+public class SearchPage {
+    public static void main(String[] args) {
+
+    }
+}
