+++ src/main/java/com/TextIt/UI/SettingsPage.java
@@ -1,0 +1,8 @@
+package com.TextIt.UI;
+
+public class SettingsPage {
+
+    public static void main(String[] args) {
+
+    }
+}
